requirements.txt
openai
python-dotenv
youtube-transcript-api
